+++ Sources/SeriousSam/ArrowDir.h
@@ -1,0 +1,18 @@
+/* Copyright (c) 2002-2012 Croteam Ltd. All rights reserved. */
+
+#ifndef SE_INCL_ARROWDIR_H
+#define SE_INCL_ARROWDIR_H
+#ifdef PRAGMA_ONCE
+  #pragma once
+#endif
+
+enum ArrowDir {
+	AD_NONE,
+	AD_UP,
+	AD_DOWN,
+	AD_LEFT,
+	AD_RIGHT,
+};
+
+
+#endif  /* include-once check. */
